:recycle: Refactor: Accordion 삭제 후 radix-ui 사용하여 Category 구현 변경
