+++ src/components/common/Navigation/SideBar/Category.stories.tsx
@@ -1,0 +1,12 @@
+import type { ComponentMeta, ComponentStory } from "@storybook/react";
+
+import { Category } from "./Category";
+
+export default {
+  title: "components/common/Accordion",
+  component: Category,
+} as ComponentMeta<typeof Category>;
+
+const Template: ComponentStory<typeof Category> = () => <Category />;
+
+export const Default = Template.bind({});
