+++ src/components/common/Navigation/SideBar/Category.tsx
@@ -1,8 +1,8 @@
+import { Content, Header, Item, Root, Trigger } from "@radix-ui/react-accordion";
 import { useRouter } from "next/router";
 
 import { useGetCategoryWithTag } from "@/application/hooks";
 import { PATH } from "@/application/util";
-import { Accordion } from "@/components/common/Accordion";
 import { useSetDrawerContext } from "@/components/common/Drawer";
 import { Icon } from "@/components/common/Icon";
 
@@ -14,15 +14,13 @@ export const Category = () => {
       const favoriteItem = {
         name: FAVORITE_ID,
         id: FAVORITE_ID,
-        children: categories
-          .map((category) => category.tags.filter((tag) => tag.isFav).map((tag) => tag.name))
-          .flat(),
+        tags: categories.map((category) => category.tags.filter((tag) => tag.isFav)).flat(),
       };
 
       const restItem = categories.map((category) => ({
         name: category.name,
         id: String(category.categoryId),
-        children: category.tags.filter((tag) => !tag.isFav).map((tag) => tag.name),
+        tags: category.tags.filter((tag) => !tag.isFav),
       }));
 
       return [favoriteItem, ...restItem];
@@ -37,28 +35,46 @@ export const Category = () => {
     router.push(PATH.getExploreByTagPath(tagName));
   };
 
-  const handleDeleteItem = (tagName: string) => {
+  const handleDeleteItem = (tagId: number) => {
     // TODO mutation
   };
 
   return (
-    <Accordion
-      defaultValue={FAVORITE_ID}
-      items={data}
-      render={(item) => (
-        <ul className="flex flex-col px-50 font-suit text-16-semibold-140">
-          {item.children.map((child) => (
-            <li className="flex w-fit gap-6 py-8 [&>#remove]:hover:visible" key={child}>
-              <button onClick={() => onClickItem(child)}>{child}</button>
-              {item.id === FAVORITE_ID && (
-                <button className="invisible" id="remove" onClick={() => handleDeleteItem(child)}>
-                  <Icon height={24} name="cancel" width={24} />
-                </button>
-              )}
-            </li>
-          ))}
-        </ul>
-      )}
-    />
+    <Root collapsible className="w-full min-w-300" defaultValue={FAVORITE_ID} type="single">
+      {data.map((item) => (
+        <Item key={item.id} value={item.id}>
+          <Header className="py-4">
+            <Trigger className="flex w-full items-center justify-between gap-8 rounded-full px-16 py-12 text-16-semibold-130 hover:bg-gray-100 data-[state=open]:bg-gray-100 [&>#chevronDown]:data-[state=open]:rotate-180">
+              <div className="h-24 w-24 rounded-full bg-light-gray-30"></div>
+              <span className="flex-grow text-left">{item.name}</span>
+              <Icon
+                aria-hidden
+                className="transition-transform duration-300 ease-[cubic-bezier(0.87,0,0.13,1)]"
+                id="chevronDown"
+                name="chevronDown"
+              />
+            </Trigger>
+          </Header>
+          <Content className="overflow-hidden data-[state=open]:animate-slide-down data-[state=closed]:animate-slide-up">
+            <ul className="flex flex-col px-50 font-suit text-16-semibold-140">
+              {item.tags.map((tag) => (
+                <li className="flex w-fit gap-6 py-8 [&>#remove]:hover:visible" key={tag.tagId}>
+                  <button onClick={() => onClickItem(tag.name)}>{tag.name}</button>
+                  {item.id === FAVORITE_ID && (
+                    <button
+                      className="invisible"
+                      id="remove"
+                      onClick={() => handleDeleteItem(tag.tagId)}
+                    >
+                      <Icon height={24} name="cancel" width={24} />
+                    </button>
+                  )}
+                </li>
+              ))}
+            </ul>
+          </Content>
+        </Item>
+      ))}
+    </Root>
   );
 };
